requirements.txt
Django==5.2.7
djangorestframework==3.16.1
easyocr==1.7.2
numpy==2.3.4
paddleocr==3.3.0
Pillow==12.0.0
scikit_learn==1.7.2
torch==2.9.0
torchvision==0.24.0
